requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
akro==0.0.8
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
better-exceptions==0.3.3
bleach==6.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==1.3.0
cma==3.3.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.35
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
dm-control==1.0.5
dm-env==1.5
dm-tree==0.1.7
docker-pycreds==0.4.0
exceptiongroup==1.1.1
executing==1.2.0
fasteners==0.18
fastjsonschema==2.17.1
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
fqdn==1.5.1
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.9
google-auth==2.19.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.2
gym==0.17.2
gym-notices==0.0.8
h5py==3.8.0
idna==3.4
imageio==2.31.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-vim==0.16.0
jupyterlab_server==2.22.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
labmaze==1.0.6
lazy_loader==0.2
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
moviepy==1.0.3
mpmath==1.3.0
mujoco==2.2.1
mujoco-py==2.1.2.14
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
numpy-quaternion==2022.4.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
proglog==0.1.10
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.19.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyglet==1.5.0
Pygments==2.15.1
PyOpenGL==3.1.7
pyparsing==2.4.7
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.25.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow-estimator==2.10.0
tensorflow==2.10.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.16
wandb==0.15.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.3.4
wrapt==1.15.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.15.0

pybullet